plugins/rune/scripts/figma-to-react/requirements.txt
# Figma-to-React MCP Server dependencies
mcp[cli]>=1.2.0
httpx>=0.27.0
pydantic>=2.0
pytest-asyncio>=0.23.0
